+++ src/test/java/voterAccess/MainControllerTest.java
@@ -2,9 +2,7 @@ package voterAccess;
 
 
 import static org.hamcrest.Matchers.equalTo;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -30,7 +28,8 @@ import org.springframework.web.client.RestTemplate;
 import org.springframework.web.context.WebApplicationContext;
 
 import es.uniovi.asw.Application;
-import es.uniovi.asw.types.UserInfo;
+
+
 
 
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -46,9 +45,9 @@ public class MainControllerTest {
 	private RestTemplate template;
 	private MockMvc mvc;
 	boolean porPantalla=false;
-	
 
-	
+
+
 
 	@SuppressWarnings("static-access")
 	@Before
@@ -56,9 +55,9 @@ public class MainControllerTest {
 		this.base = new URL("http://localhost:" + port + "/");
 		template = new TestRestTemplate();		
 		mvc = new MockMvcBuilders().webAppContextSetup(wac).build();
-		
+
 	}
-	
+
 	/*
 	@Test
 	public void getUser() throws Exception {
@@ -73,7 +72,7 @@ public class MainControllerTest {
 		assertThat(response.getStatusCode(), equalTo(Status.OK));
 		assertThat(response.getBody(), equalTo(expected));
 	}*/
-	
+
 	/*
 	 * 
 	 * 
@@ -89,35 +88,38 @@ public class MainControllerTest {
 		assertThat(response.hasBody(), equalTo(true));
 		if (porPantalla) { System.out.println(response.getBody()); }
 	}
-	
+
 	@Test	
 	public void printScreen() throws Exception {
 		MvcResult m = (mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
-			
+
 				.content("{\"login\":\"user1@mail.com\", \"password\": \"user1\"}")
-		).andReturn());
-		System.out.println("RESULTADO JSON OBTENIDO: "+ m.getResponse().getContentAsString());
+				).andReturn());
+
+		if (porPantalla){ 
+			System.out.println("RESULTADO JSON OBTENIDO: "+ m.getResponse().getContentAsString());
+		}
 	}
-	
+
 	@Test
 	public void postUserOK() throws Exception {
 		mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
-				.content("{\"login\":\"user1@mail.com\", \"password\": \"user1\"}"))
+				.content("{\"login\":\"user2@mail.com\", \"password\": \"user2\"}"))
 		.andExpect(status().isOk());
 	}
-	
+
 	@Test
 	public void postUserKnown() throws Exception {
+		//Voter vo = new Voter(nombre, email, password, nif, pollingStationCode)
+		//Voter voter1 = new Voter("User1", "user1@mail.com", "user1", "12312321Q", 123);
 		
 		mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
-				.content("{\"login\":\"user1@mail.com\", \"password\": \"user1\"}")
-		).andExpect(status().isOk())
-		 .andExpect(jsonPath("name", equalTo("User1")));
-		 //.andExpect(jsonPath("pollingStationCode", equalTo("123")));
-		 //.andExpect(jsonPath("pollingStationCode").value("123"));
+				.content("{\"login\":\"user2@mail.com\", \"password\": \"user2\"}")
+				).andExpect(status().isOk())
+		.andExpect(jsonPath("name", equalTo("User2")));
 	}
 
 	@Test
@@ -125,113 +127,91 @@ public class MainControllerTest {
 		mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content("{\"login\":\"emailnovalido666@mail.com\", \"password\": \"password\"}")
-		).andExpect(status().isNotFound());		 
+				).andExpect(status().isNotFound());
 	}
-	
+
 	@Test
 	public void postUserBadPass() throws Exception {		
 		mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content("{\"login\":\"user1@mail.com\", \"password\": \"pas\"}")
-		).andExpect(status().isNotFound());		 
+				).andExpect(status().isNotFound());		 
 	}
-	
+
 	@Test
 	public void postUserInvalidMailFormat_at() throws Exception {		
 		mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content("{\"login\":\"user1mail.com\", \"password\": \"user1\"}")
-		).andExpect(status().isNotFound());		 	
+				).andExpect(status().isNotFound());		 	
 	}
-	
+
 	@Test
 	public void postUserInvalidMailFormat_domain() throws Exception {		
 		mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content("{\"login\":\"user1@gmail.com\", \"password\": \"user1\"}")
-		).andExpect(status().isNotFound());			
+				).andExpect(status().isNotFound());			
 	}
-	
+
 	@Test
 	public void postUserInvalidMailFormatDomain_wrongPass() throws Exception {		
 		mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content("{\"login\":\"user1@mail.com\", \"password\": \"pa\"}")
-		).andExpect(status().isNotFound());		
+				).andExpect(status().isNotFound());		
 	}
-	
+
 	// COMPROBACION DE CAMBIO DE PASS 
-	
-	
-	
+	@Test
 	public void postUserChangePass() throws Exception {		
 		//cambia pass
 		mvc.perform(post("/ChangePassword")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content("{\"login\":\"user1@mail.com\", \"oldPassword\": \"user1\", \"newPassword\": \"newPass\"}")
-		).andExpect(status().isOk());		
-				
+				).andExpect(status().isOk());		
+
 		//comprueba pass nueva
 		mvc.perform(post("/user")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content("{\"login\":\"user1@mail.com\", \"password\": \"newPass\"}")
-		).andExpect(status().isOk())
-		 .andExpect(jsonPath("name", equalTo("User1")));
-		 //.andExpect(jsonPath("pollingStationCode", equalTo("123")));
+				).andExpect(status().isOk())
+		.andExpect(jsonPath("name", equalTo("User1")));
+		//.andExpect(jsonPath("pollingStationCode", equalTo("123")));
+	}
+
+	@Test
+	public void postUserChangePassErrorUser() throws Exception {		
+	
+		mvc.perform(post("/ChangePassword")
+				.contentType(MediaType.APPLICATION_JSON)
+				.content("{\"login\":\"user1@mail.com\", \"oldPassword\": \"user1\", \"newPassword\": \"newPass\"}"))
+			.andExpect(status().isOk());		
+
+		mvc.perform(post("/user")
+				.contentType(MediaType.APPLICATION_JSON)
+				.content("null"))
+			.andExpect(status().is4xxClientError());
+	}
+
+	@Test
+	public void postUserChangePassErrorChangePass() throws Exception {		
+	
+		mvc.perform(post("/ChangePassword")
+				.contentType(MediaType.APPLICATION_JSON)
+				.content("null")
+				).andExpect(status().is4xxClientError());	
+
 		
-	}
+		mvc.perform(post("/user")
+				.contentType(MediaType.APPLICATION_JSON)
+				.content("{\"login\":\"user1@mail.com\", \"password\": \"newPass\"}")
+				).andExpect(status().isOk())
+		.andExpect(jsonPath("name", equalTo("User1")));
 	
-
-	/*
-	 * 
-	 * BLOQUE DE TESTS DEDICADOS AL: DOMINIO
-	 * 
-	 */
-
-
-	
-	//UserInfo
-	@Test
-	public void testUserInfoVacio() throws Exception {
-
-		UserInfo ui = new UserInfo();
-		assertTrue(ui.hashCode()==923521);
-		assertTrue(ui.toString().equals("UserInfo [name=null, "
-				+ "nif=null, "
-				+ "email=null, "
-				+ "pollingStationCode=null]"));
 	}
-	
-	@Test
-	public void testUserInfoCompare() throws Exception {
 
-		UserInfo ui1 = new UserInfo();
-		UserInfo ui2 = new UserInfo();
-		//compara 2 null
-		assertTrue(ui1.equals(ui2));
-		ui2 = new UserInfo("name", "nif", "email", 123L);
-		// compara un null con un bien formado
-		assertFalse(ui1.equals(ui2));
-		// comapra 2 bien formados
-		ui1 = new UserInfo("name", "nif", "email", 123L);
-		assertTrue(ui1.equals(ui2));
-		//compara dif nif
-		ui1 = new UserInfo("pepe", "25", "email@dom.com", 1L);
-		ui2 = new UserInfo("pepe", "50", "email@dom.com", 1L);
-		assertFalse(ui1.equals(ui2));
-		//compara dif nombre
-		ui1 = new UserInfo("pepe", "25", "email@dom.com", 1L);
-		ui2 = new UserInfo("juan", "25", "email@dom.com", 1L);
-		assertFalse(ui1.equals(ui2));
-		//compara dif colegio electoral
-		ui1 = new UserInfo("pepe", "50", "email@dom.com", 8L);
-		ui2 = new UserInfo("pepe", "50", "email@dom.com", 1L);
-		assertFalse(ui1.equals(ui2));
-		//compara dif correo
-		ui1 = new UserInfo("pepe", "50", "mail@dom.com", 1L);
-		ui2 = new UserInfo("pepe", "50", "email@hotmail.com", 1L);
-		assertFalse(ui1.equals(ui2));
-	}
+
 	
 
 }
